html/semantics/embedded-content-0/media-elements/loading-the-media-resource/035.html

<!doctype html>
<title>await a stable state and sync event handlers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<video></video>
<script>
var v;
var t = async_test(function(t) {
  v = document.querySelector('video');
  var a = document.createElement('a');
  a.onclick = function() {
    v.setAttribute('src', '#'); // invokes media load which invokes resource selection
  };
  a.click(); // sync fires click, so sets src
  // now we should still await a stable state because the script hasn't
  // finished, the event handler has just returned
  v.removeAttribute('src');
});
</script>
<script>
t.step(function() {
  // now the sync section of resource selection should have run and should
  // have found no src="" or <source> thus networkState being set to 0.
  // if the sync section was run when onclick returned, then networkState
  // would be either 2 or 3.
  assert_equals(v.networkState, 0);
  t.done();
});
</script>
